First stage of switching order of args to subtype functions
Summary:
The order of arguments to sub_type and associated functions in Typing_subtype is the
source of much confusion and head-scratching. This is the first stage of switching
the arguments to sub, super which is the order in which they are usually presented
e.g. in notation such as sub <: super.

The first change just changes the internal functions but not the main exported one.
The second change affects the exported functions and their use in typing.ml etc.

In a separate diff, we will switch around the pattern matching in Typing_subtype so
that it matches the order of arguments.

Reviewed By: ericlippert

Differential Revision: D3503600

fbshipit-source-id: 34bb282e6356b82ecc7265fa7fadd0f2dbe8869a