Measurement and logging of size of deserialized values
Summary:
The sharedmem_log_level=1 setting currently records how many bytes a value
occupies when in the shared heap. We now add stats for the size of the value
when it is deserialized into the Caml GC heap.

Note: the function HeapSize.size might be useful elsewhere; this is just an
example that's interesting because we can compare the size of values in
serialized and deserialized form.

Reviewed By: dabek

Differential Revision: D3514884

fbshipit-source-id: cc9d0197e747370bb2ab19743782f35b28a054e8