Delete StructArray and Shape
Summary:
We're not using them, and the sentiment I've heard from the team is
that if we want them we're going to have to rewrite them from scratch.

This still leaves behind the NewStructArray opcode.  The last time I tried to
delete that it was a performance regression, presumably because of how the
emitter packs the array element pairs more efficiently.

Reviewed By: paulbiss

Differential Revision: D3505834

fbshipit-source-id: c7dda943f93d2e2026d205b52490104240155f6c