Fix typo bug in annotCompat()
Summary:
When KindOfPersistentArray was added, a change to add the usage of isArrayType()
in annotCompat() apparently went wrong. Instead of isArrayType(dt),
isArrayType(KindOfArray) was added instead, which of course is always true. As a
result, the compiler optimizes away the expression entirely. Fix it to be
isArrayType(dt).

Fover ok:

I noticed this as part of the Hack array changes and fixed it in that
stack. However, fixing this might be part of the perf regression, so I decided
to split out on its own. This bug has been present since November. I'm very
surprised it doesn't seem to actually break anything.

Reviewed By: markw65

Differential Revision: D3521845

fbshipit-source-id: 8126c2687f4614d05c9ae38c146b886f39de9e21