Allow string cast of float negative zero to print negative zero
Summary:
We were forcing a string cast of negative zero to print `"0"`, but PHP5 and PHP7 both render that as `"-0"`. We should be consistent when HipHopSyntax is off. (FB relies on the other behaviour, and from a user perspective it makes a bit more sense anyway IMO.)

Fixes https://github.com/facebook/hhvm/issues/7156

Reviewed By: paulbiss

Differential Revision: D3522010

fbshipit-source-id: 77aabccd7e411cd012f65ef55d15e2975f8aad6b