Break dependence of SOAP extension on DataType values
Summary:
The SOAP extension defines various integer values to represent types
internally. However, these values are defined as a super-set of our DataType
values, so that it can use Variant::getType() to directly map to one of its
equivalent types. This is annoying because it means if we adjust the DataType
values, we might have to likewise adjust the SOAP type values (which are visible
in PHP). Break this dependence by adding a mapping function which maps DataType
values to their equivalent SOAP values, and use that where-ever needed.

Reviewed By: edwinsmith

Differential Revision: D3515731

fbshipit-source-id: c3a17f5f897a84b9085f4d48d418d81a33df1886