Avoid infinite loops in type-scanners more efficiently
Summary:
Normally the auto-generated type-scanners do not have to worry about entering
into cycles. This is because they, by default, never follow pointers, and thus
cannot enter into a cycle. However, there are certain constructs, which despite
containing pointers, we want to treat as flat values. Standard library
containers are one such example. Since it is possible to enter into a cycle with
these, a runtime check needs to be done to avoid looping infinitely.

The initial implementation of this check just used a set of visited pointers,
which was consulted before scanning a pointer. The problem with this is that
very large data structures can cause the visited set to grow very large, and
clearing the set after the scan can consume a large amount of time.

The visited set has the nice property that we'll never visit a node more than
once, but this is strictly unnecessary. The only real requirement is that we
don't loop infinitely. It is merely undesirable that work isn't duplicated. For
this reason, replace the visited set with a stack of pointers currently being
visited. Push a pointer onto the stack when scanning it, and pop it off when
done. If a pointer is already on the stack, do not scan it. Since chains of
scanned pointers tends to low, linearly searching the stack is sufficient.

Reviewed By: edwinsmith

Differential Revision: D3521753

fbshipit-source-id: e50abcd8b34a3a73c43c58f08055d3b63c685c13