Fix hh_server on OSX
Summary:
On OSX, we use `shm_open` to create a file descriptor for the shared
memory. Unlike the Linux code, this creates a file descriptor with the `FD_CLOEXEC`
(close-on-exec) flag set. Due to the complicated way that hack sets up the
shared memory across multiple processes, something must be closing the file
descriptor.

Anyway, this code fixes. Tested on Linux and OSX.

Reviewed By: dabek

Differential Revision: D3530260

fbshipit-source-id: bc7c69cca62ea1b3f1acc2c66aea86b0825f448a