Fix pretty-parser tests to not track comments
Summary:
The ffp tests parse, pretty-print, parse again and compare, ignoring whitespace.
But it did not compare ignoring comments. In a couple cases we had:

$x =
/* comment */
1;

being pretty-printed as "$x = /* comment */ 1;" , which moves the ownership of the comment from the 1 to the =. The test was then failing.

I've modified the tests so that they strip comments before comparing pretty-printed parse tree to original parse tree.

Differential Revision: D3531492

fbshipit-source-id: 2b714c682d11539d5b5a90f8d509735ee393ae3d