Check the Phar has metadata before attempting to unserialize() it
Summary:
Phar metadata is optional with the metadata length set to 0 if not present.  This confirms patch checks that the metadata length is greater than 0 before attempting to unserialize() it.
- https://secure.php.net/manual/en/phar.fileformat.phar.php

This is in response to the listed issue in #7178, however there are other outstanding/unrelated type checking warnings when running in PHP7 scalar mode (null instead of string).
Closes https://github.com/facebook/hhvm/pull/7197

Reviewed By: Orvid

Differential Revision: D3510970

fbshipit-source-id: 435376150c0245187092c47bdd08925d6ba0ffca