Fix sharedMem dependency while parsing
Summary:
We call lint.add in the parser because "some things that we want to lint are discarded [after the parsing phase]", however we should only be adding things to this list if we're planning on linting, which means only if lint_list has been initialized.

This diff switches the order of checking fixmes and lint_list to avoid a dependency on initializing sharedMem if we are not linting.

Reviewed By: ericlippert

Differential Revision: D3542534

fbshipit-source-id: a9a861326b852476ba9272d9d0c16296a9f61e47