Restore --remove-dead-fixme option with better help comment
Summary:
Having investigated the problems associated with `--remove-dead-fixme` option in production it is clear that
it works correctly if (a) mini-state loading is not happening and (b) the server starts clean. Both can
be ensured by running
```
hh restart --no-load
```
prior to running
```
hh --remove-dead-fixme <code>
```
I've updated the help text on `hh_client` to reflect this.

Making `--remove-dead-fixme` work in incremental mode and with mini-state loading is a non-trivial amount of
work, as we would need to save a lot more state. The feature is sufficiently useful in its current form that
I think it's worth having available.

Reviewed By: dlreeves

Differential Revision: D3542535

fbshipit-source-id: 8d685b0667dfad0421c8ab399283a8bdf6c12d51