Improved parse tree traversal
Summary:
I'm going to be improving the error reporting post-parse pass soon.

Some errors will require context; for example, we need to know whether a formal parameter that is missing a type annotation is in a lambda or not.

The existing error pass only examines individual nodes, rather than the node in the context of its parents.

I've added two folding methods which take a function from accumulator, node, parent-list to new value of accumulator.

The improved error reporting code will come in a later diff.

Differential Revision: D3544048

fbshipit-source-id: 45d9623e40c202498d24cdce698bdea2f8efadd0