Export more stats about performance of IO Thread
Summary:
Now exporting: idle time, busy time, notification queue
Snapshots of the data are as before, being captured at the end of operation and is attached to result.
idle and busy look very weird (with negative values) in the beginning of execution. So I'm not sure how they will look like in a long running process

Reviewed By: abal147

Differential Revision: D3478508

fbshipit-source-id: e2b1fc6ed8346f6e3743df04df9c86907d3dc4dc