Fix async stepping for non-blocked async operation
Summary:
The new async stepper allows "await" opcode to pass through the flow filter. However if the async operation is non-blocking we return AsyncStepHandleOpcodeResult::Unhandled, this caused normal non-async stepping code to handle it and result in break on the "await" opcode which is incorrect.
Fix it by returning AsyncStepHandleOpcodeResult::Handled so that AsyncStepper is in charge of the stepping even in non-blocking mode.
Testcase will be added in later diff.

Reviewed By: johnislarry

Differential Revision: D3544280

fbshipit-source-id: f5e64281004ee140c1cae8ccda23c74db2d86af7