Allow Closures to be inlined
Summary:
We didn't explicitly prevent them being inlined, but they usually
TRACE_PUNTed because the first local was marked as Uninit, where it
should actually be the Closure object (and there are AssertTypes to
that effect). We also didn't set things up properly if we /did/ manage
to inline them, so fix that too.

Also noticed that we put the empty Variadic Param in the wrong slot
(the last local, rather than the last param), which also tended to
cause functions with variadic args to not be inlined due to
TRACE_PUNTs, so fixed that too.

Reviewed By: paulbiss

Differential Revision: D3552782

fbshipit-source-id: d98bb914c750b4a237ac91b0af74cbabe12c49aa