Add error for break and continue in wrong places
Summary:
We now detect when a break or continue statement appears in an illegal location.

The test is simple: we must find a parent that is a loop or switch (in the case of break) before we find an enclosing anonymous method.

I've also done some small refactoring of the error reporting code. Error reporting for statements is consolidated, for instance.

Also, errors reported are now sorted by offset in the file; this way we can change the order in which errors are found without impacting the test case baseline.

Differential Revision: D3557575

fbshipit-source-id: 20bddef3fbc213ba35bc2a608b07959ebdba02b6