fix enum parsing
Summary:
The old parser requires that the last enumerator in the enum declaration also ends with a semicolon. Our new parser requires the opposite, i.e. the last enumerator does not end with a semicolon. This diff changes the behavior to what the old parser does.
New grammar:
enumerator-list :
	enumerator-list enumerator
enumerator:
	enumerator-constant = constant-expression ;
The other parts are not changed.

Test cases are updated to reflect this change.

Reviewed By: ericlippert

Differential Revision: D3556657

fbshipit-source-id: 7788dee52c2ba08e784a87abeac2231159ceecdd