Optimize fixupWork and CallerFrame
Summary:
Many uses of CallerFrame only needed the ActRec, and didn't need to
clean the vm state, so add an optimized callerFrameHelper to do just
that.

The only hot use of isVMFrame() was from FixupMap::fixup() (and the
new callerFrameHelper), so move it to fixup.cpp where it can be
inlined (I didn't want to add more includes to bytecode.h).

While there, optimize the hot loop in fixup; ar == ar->m_sfp should
never be true (ActRec::sfp() used to return 'this' if we were the last
vm frame, but it returns nullptr now - and we don't call that anyway),
and we already crashed if rbp == nullptr, so just remove the condition,
and loop forever.

Reviewed By: mxw

Differential Revision: D3573308

fbshipit-source-id: 92ef209c40ed4ea3a4e200a954c9a4b385299651