Enable scribe if gc is enabled, log cmdline args
Summary:
Add support for cmdline logging to scuba when GC is enabled.
Capture globals["argv"] as a normvector (array of strings)
(for any table).

Adjusted the logic to enable gc sampling: wait until the
first collection of each request to decide whether to do
logging for that request, to make sure that a) runtime options
are set and b) StructuredLog is initialized, before setting
t_enable_sampling.

Reviewed By: kodafb

Differential Revision: D3557411

fbshipit-source-id: 8f382bb1701754b47a5ab67d9de4955d7a06b701