Support php_uname on Windows
Summary:
What, you liked your simple implementation? Well, you can't haz.

This pulls in the php_uname implementation for windows from PHP, with modifications to reflect the existence of C++.
This was the only place that was left that was using `uname`, and my portability implementation of that was terrible, so implement PHP's version correctly here.

Reviewed By: paulbiss

Differential Revision: D3014097

fbshipit-source-id: c4fea2d9d225090b5e4d699a438ea3ea5c470fcf