Clean up error reporting code
Summary:
I've moved all the code which expects a particular token and gives a specific error into helper methods.

This has several nice properties.

First, it emphasizes the business domain of the parser by hiding the error reporting mechanism.

Second, it means that there will be fewer places to update when eventually we reconcile my quick-and-dirty error reporting with the actual Hack error reporting mechanisms.

Third, it removes a source of cut-n-paste bugs by ensuring that the same error is always reported for the same missing token.

Differential Revision: D3583107

fbshipit-source-id: 744b68250e7ad533b1541f9e2057eb7844812e8f