Fix irgen::ifElse() when the branch instruction is elided
Summary:
If the branch instruction that should terminate `branch_block' is
optimized away, we shouldn't be trying to save its out state, since
we're going to emit (at least part of) `next' into the same block.

To rectify this, if `branch_block' would fallthru to next, instead
add a Jmp to an explicitly constructed `next_block'.  This will get
optimized away eventually, and simplifies the logic in ifElse().  It
also makes it analogous to ifThen(), so share the code.

Reviewed By: markw65

Differential Revision: D3586775

fbshipit-source-id: 4834d7c4c1e5c7c5d7eb3201b82291be10588d55