Factor out function header
Summary:
Function header is useful in itself because we have abstract methods that do not have a body. Thus I think it makes sense to create a node for function header.

Eric also observes that attribute specification is not necessarily part of function header since it is always in front and thus it is taken out of function header.

Differential Revision: D3564135

fbshipit-source-id: fcc09af4984eb3699979765162db32112767a5ba