Replace Process::GetProcessId with getpid()
Summary:
Now that we have folly/portability/Unistd.h, we can use getpid() on all
platforms.

For the cases where we want to print the pid_t, we'll just use a C-style cast to
int64_t.  This way it will work even for platforms that have pid_t as a pointer.

Also get the single call site of Process::GetParentProcessId() to use getppid().

Reviewed By: Orvid

Differential Revision: D3587574

fbshipit-source-id: ab57017c3431e00fec083fcd32029974d6e2b045