Jury-rig a solution to spurious (incorrect) results in "go to definition"
Summary:
This is an ugly-ish Hack that ads a dependency on the path that positions coming from buffer will be at to use it for filtering the results that we don't want.

In the ideal world, taccess_hooks would be called only on type const accesses in the file that is being currently typechecked. But the place where it is dispatched (see {D3203914}) is also called when resolving types coming from other files.

Instead, it probably should be dispatched from decl_hint, similarly to how class_id hook is dispatched: diffusion/FBS/browse/master/fbcode/hphp/hack/src/decl/decl_hint.ml$72?view=highlighted

But in that phase of checking, it's relatively hard to resolve that "C" in a typehint like A::B::C is actually a reference to D::C - the only place where it is resolved is deep inside Typing_access (where the hook is dispatched right now), and it's rather tightly coupled with typing phase.

Reviewed By: andrewjkennedy

Differential Revision: D3522307

fbshipit-source-id: 8622c24eaa17e58fe4cebe571dc3900d418d8f2e