second pass error reporting on method declarations
Summary:
Report errors on malformed method, constructor, destructor declarations based on the grammar specification. An interesting case which is not tests is whether all methods, constructors and destructors need to have a visibility modifier. The spec specifies that there needs to be SOME modifier but didnt specify which one, which looks fishy to me.
This is on top of the changes I accidentally commited in diff D3564210
More fine-grained error location reporting to come later

Reviewed By: ericlippert

Differential Revision: D3591180

fbshipit-source-id: 5f72b85ab3982aae338efe30f15f61efd8953720