hack grammar descriptor
Summary:
This is a description of the hack language, for generating random ASTs.
This grammar described here does not strictly follow the grammar descibed in github, because:
1. The grammar described in github is not entirely correct.
2. Some nonterminals are pre-expanded to make the grammar more manageable
3. Since the grammar in github is not entirely correct, some derivations are added/deleted to my best knowledge

The grammar is incomplete now, it contains all the rules that the current full fidelity parser can parse. This does not mean that all the strings that can be parsed will be generated, but it should be close. As the parser is being completed, more derivations will be "unlocked" (as they are commented out now).

Reviewed By: ericlippert

Differential Revision: D3441711

fbshipit-source-id: 4bafdb301fad4df4ff0bc5dc643994fc895aaf52