Make extension configuration occur before anything else
Summary:
This is needed for two reasons.

The first is that, with the way it is now, the `ENABLE_EXTENSION_*` defines are only valid in the `hphp/runtime/ext` directory. This means that it's currently impossible for me to disable the `ext_process` extension, as is required on windows (php disables it on windows as well), as it requires disabling a hook used in the runtime. By moving the configuration up, it makes it a simple matter of surrounding the effected lines of code with `#ifdef ENABLE_EXTENSION_PROCESS`.

Reviewed By: paulbiss

Differential Revision: D2881805

Pulled By: Orvid

fbshipit-source-id: f709fbd209467b06a87111d250204537162c1884