Avoid copying xdebug messages
Summary:
Each xdebug_xml_node object passed to XDebugServer::sendMessage() is first
turned into a String, and then that String is appended to a StringBuffer, making
another copy of the string.  These strings can be really large, so avoid using
StringBuffer by doing a writev() with the few buffers we want to send along the
socket.

Reviewed By: yinghuitan

Differential Revision: D3615117

fbshipit-source-id: b354f865ac34245ed4735955826781303a9eba85