Allow qualified names that omit '\' as a prefix
Summary:
This solution is straight-forward and it seems to work, but it's not particularly elegant, so open to suggestions.

We were parsing these correctly but also counting them as errors. An example is
hphp/hack/test/facebook/ffp/tests/namespace_async_2.php which had parser error
(14,5)-(14,8) An unqualified name is expected here.
due to the inclusion of an error when detecting a '\' character in a qualified name

Reviewed By: ericlippert

Differential Revision: D3609582

fbshipit-source-id: 2b6c312a0bf46b46cf6bc26286ffe64d661e561e