Fix various pseudoMain issues
Summary:
Various parts of the jit assume that the m_this field of an ActRec
must match the m_cls field of the Func. We try to ensure this by
cloning pseudoMains to match the calling environment. But there were
places where we called the null-class pseudoMain with a non-null $this
or a non-null caller-context.

This fixes the callsites, and removes the default nullptr from getMain
so that writers of new code will have to at least think about what
they're passing in.

Reviewed By: paulbiss

Differential Revision: D3615279

fbshipit-source-id: 6bbd7a2f19581f1799ea62d90a738ec4f5f017ca