Fix parsing catch clauses
Summary:
The specification says that a catch clause has the form

    catch(parameter-list) { ... }

such that the list has exactly one element. But this is wrong. First, it seems silly
to say that we have a list with exactly one element; why not just say that we have
an element? But more seriously, none of the other productions that are legal for a parameter
list with one element are legal here.  "<<foo>> Exeception $x = null," is a legal one-element
parameter list, but a catch clause may not have an attribute, a default value or a trailing comma.

Trying to characterize this thing as a list is silly. I've changed the spec grammar to

    catch(type-specifier variable-name) { ... }

and updated the parser to match the new spec.

Reviewed By: eshrews

Differential Revision: D3614736

fbshipit-source-id: 044e8a609f7f2f7faccd4c308cd56a7da2620d3c