Add error detecting misused ellipses in parameter lists.
Summary:
We give two separate errors for (1) an ellipsis appears before the end of a formal parameter list, and (2)
an ellipsis appears at the end of a formal parameter list followed by a comma.

It might be overkill to separate out these two cases; we could consider combining them into one when we refactor the error reporting code.

Note however that in the first case the error is reported at the location of the ellipsis; the second is reported on the comma. Subtle.

Reviewed By: eshrews

Differential Revision: D3617184

fbshipit-source-id: 8551a8be9d428dee06acf25f3fffade824a5d034