Reverted commit D2881805
Summary:
This is needed for two reasons.

The first is that, with the way it is now, the `ENABLE_EXTENSION_*` defines are only valid in the `hphp/runtime/ext` directory. This means that it's currently impossible for me to disable the `ext_process` extension, as is required on windows (php disables it on windows as well), as it requires disabling a hook used in the runtime. By moving the configuration up, it makes it a simple matter of surrounding the effected lines of code with `#ifdef ENABLE_EXTENSION_PROCESS`.

Reviewed By: paulbiss

Differential Revision: D2881805

Pulled By: paulbiss

fbshipit-source-id: 6483c97d3ea24cdafbc2d54f111e3df12ed721e0