Add new check for incompatible enum types
Summary:
Currently it's possible to use `===` equality on two enum types that
are distinct and abstract (have no "as" clause). It can be argued that
this is usually a source of potential bugs, and so this diff introduces
a new error for this situation. It's always possible to override by
casting up to `mixed`.

Reviewed By: eshrews

Differential Revision: D3613038

fbshipit-source-id: be468a5332940e38b8ebbdb8c546df638f45ed08