Have decl_class_type replace class_type in the shared heap
Summary:
Now that there is a Decl_heap and a Typing_heap, we are free to change what we store in Decl_heap. Then in the typing heap we only need a way of reconstructing the typing phases view of the shared heap data by transforming the data we store in the Decl_heap. It illustrate this we change Decl_heap.Classes to store the `decl_class_type`. This is basically an exact copy of `class_type` but with different names for the fields. Later on there will be more of a meaningful difference, but this will be the start.

Depends on D3539297

Reviewed By: andrewjkennedy

Differential Revision: D3539453

fbshipit-source-id: 63df1d25eabf69af6cc74ef597ddd0334e8766e2