Delay substitution of type parameters when inheriting members
Summary:
Consider:

```
class Base<T> {
  public function do(T $x): void {}
}

class Child<T as arraykey> extends Base<T> {}

final class FinalChild extends Child<T> {}
```

Currently we would fold the members `Base::do` in each class and substitute any relevant type parameters a long the way. In the end we will have stored roughly the following types in shared memory

```
Base::do       -> (function (T $x): void)
Child::do      -> (function (T $x): void) with T as arraykey
FinalChild::do -> (function (int $x): void)
```

As you can see a lot of data is repeated in the type each time we fold the member. Instead of folding the complete type, it would be more efficient to say that a particular child references a parent's type. i.e.

```
Base::do       -> (function (T $x): void)
Child::do      -> Base::do with T as arraykey
FinalChild::do -> Base::do with T = int
```

The only difference between `Base::do` and `Child::do` is the fact that type parameters has been substituted. We can delay this substitution after {D2613160} because the extra scoping allows us to disambiguate type parameters. That is what this diff does.

When we are inheriting members we build up a substitution table for type parameters. We store this table as part of the `decl_class_type`. We then use this table to substitute in type parameters for class members when a `decl_class_type` is converted to a `class_type`. For the above example we would have the following substitution tables

```
Base:
  {}
Child:
  { Base\T -> Child\T }
FinalChild:
  { Base\T -> Child\T, Child\T -> int }
```

The type for `Base::do` will then be `(function (Base\T $x): void)`, and the substitution table will be applied to this type in the child class.

Note:
This diff only delays the substitution. It does not take advantage of the fact now the type data is clearly repeated. Because of that this diff is a regression interms of memory and CPU time, but this will be fixed in follow up diffs

Depends on D3539453, D3561665

Reviewed By: andrewjkennedy

Differential Revision: D3561842

fbshipit-source-id: 8c0c0376fef3f46dfbf921950e5d57031d6594f0