Store the type of class elements separately in Decl_heap
Summary:
After {D3561842} we are delaying substitution of type parameters until when we translate a `decl_class_type` to a `class_type`. Now when we fold members the `ce_type` will be identical in all children when inherited from a parent class.

By far the most expensive portion of a `class_elt` is the `ce_type`, so we store it separately in the shared heap. This allows multiple classes to refer to same element. We can retrieve the type by looking at the `ce_origin` of a `class_elt`, combined with the name of the element.

To accomplish this diff does the following:

* Add `Decl_defs.element` which contains more or less the same fields as `Typing_defs.class_elt` except for the type. We include whether the element is abstract or not since it is commonly accessed and cheap to duplicate
* Store the types for properties, static properties, methods, static methods, and constructors each in `SharedMem`
* When transforming from a `Decl_defs.decl_class_type` to a `Typing_defs.class_type` we reconstruct `Typing_defs.class_elt` by using the name of the element and the origin of the element to retrieve the type

Note:
This diff is not safe to land on its own because I need to make changes to incremental mode. It is expected those tests will fail. The incremental mode changes will be a part of another diff.

This diff reduces the amount of data in shared memory, but negatively impacts cpu time. This will also be addressed in a follow up diff

Depends on D3561842

Reviewed By: andrewjkennedy

Differential Revision: D3564703

fbshipit-source-id: ae8d8aa23cdad9a954bb6e834369dabac991ab43