Lazy load type for class_elts
Summary:
This is the final diff for changing how we lay out `class_type` in shared memory. Now that the type of a `class_elt` is stored in a separate field in shared memory we can choose to load this type only when necessary. Since the type of a `class_elt` can be rather large, loading it on demand instead of premptively saves both memory and CPU time.

The easiest way I saw of doing this is making the `ce_type` field be lazy. Then in most cases we can use lazy pattern matching to force the lazy value when we need it.

Depends on D3589777

Reviewed By: andrewjkennedy

Differential Revision: D3598965

fbshipit-source-id: 2bfafb1c850e721095664bcd1d9aae63e703f2c4