Add module for signaling and reading perf_event samples
Summary:
The intended use case of util/perf-event.h is as follows:
  - Enable it for a thread on a per-request basis, with low sampling
    frequency.
  - Set a surprise flag in the signal handler whenever a sample is
    collected.
  - Consume all samples at the next surprise flag check.

Currently, the module is only set up to track memory instructions, with
the intent of profiling and categorizing HHVM's loads and stores.  It is
designed to enable examining RDS (and other thread-local memory) without
races, and to defer access to the request heap for when it's in a
consistent state, but without too much delay (so that we can make a
"best guess" of what sort of heap objects triggered a perf sample).

Some things that could be parametrized but currently aren't:
- Size of the mmap buffer.
- Events to care about.

Reviewed By: kodafb

Differential Revision: D3602779

fbshipit-source-id: ae7f15ebd07dc4df6027dce61d9b2f8dc65dd69f