Log various points along the RFH curve with StructuredLog
Summary:
We've decided that in addition to warmed up steady-state performance,
we should be focusing on RFH as well. RFH is "requests served in first hour",
which is just what it sounds like: how many requests was the server able to
serve in its first hour of uptime? This diff logs RFX for various values of X,
ranging from 30s to 2h, using StructuredLog.

Reviewed By: binliu19, kodafb

Differential Revision: D3641690

fbshipit-source-id: a00d49d78ba13e935efa32a7a643638aea708661