Fix crash with throwing from __sleep() with packed arrays in APC
Summary:
apc_store() can crash under the following conditions:

- A packed array contains an object
- That object defines the __sleep() magic method
- That magic method throws

The reason is because we pro-actively set the final count in the APCArray being
created. If an exception is thrown while populating the array, we correctly
destroy the APCArray, but since its final count is set, its destructor tries to
access uninitialized data. The solution is to set the contained count to only
the amount we succeeded in initializating before destroying the APCArray.

Reviewed By: swtaarrs

Differential Revision: D3652661

fbshipit-source-id: 351ff376ad9bd34801fd76da65c6ccc6404ab30a