Make ClassSpec trustworthy
Summary:
Often ClassSpecs were set up based on the name of a Class, without
verifying that it was unique, and then it was up to the user of the
ClassSpec to ensure that the Class* was safe to use at that point in
the program. These checks were applied pretty inconsistently.

This moves all the checks to the point where the types are setup, so
that if type().clsSpec() retuns a Class* its always safe to use it, modulo
guard relaxation.

It also replaces lookupClassOrUniqueClass with
lookupUniqueClassInContext, so that its result is always trustworthy
too.

Reviewed By: swtaarrs

Differential Revision: D3647068

fbshipit-source-id: 7fcd6a3c6969a3064abffdd2374e2013963e4e65