Enable TC dump to include TargetProfile data
Summary:
Add option to include TargetProfile data when dumping the TC.
Add toString methods for target profile classes that didn't have any.
When option is enabled,
1. invoke toString when data is gathered for optimization,
2. store this debug info in ProfData, and
3. include the data as annotations in the TC dump.

Thus, data is only included for translations actually used when optimizing a function, and reflect the state of the target profile at optimization time.

Reviewed By: paulbiss

Differential Revision: D3649521

fbshipit-source-id: 273e0e99b785ba8141cc1fb978eec36cb5778e76