Optimize instanceof static
Summary:
$x instanceof static;

was implemented as

  $x instanceof get_called_class();

We have an optimization for get_called_class which just loads the
class from the ActRec, and loads the name from that. But the emitter
generated an FPushFuncD/FCall rather than an FCallBuiltin, which
doesn't get optimized. So we ended up doing a generic call to
get_called_class(), and then a class lookup on the result.

Changing the emitter to spit out an FCallBuiltin guarantees we get the
optimization, which avoids the call, but we still did an LdClsName on the
result, and then a LookupClsRDS on the result of that. So this adds a
simplify for LookupClsRDS(LdClsName(ssa)) -> ssa

Reviewed By: jano

Differential Revision: D3672215

fbshipit-source-id: 6ecc3c061f100762e08586f42a2e80b0a4636aef