Add a new knob for controlling the amount of profiling code emitted
Summary:
We primarily use requests to control how many functions we profile.  A
"request" is not a very stable unit of PHP coverage so we can end up with
wildly varying amounts of code in code.prof and code.hot, even among servers
warming up at exactly the same time. If you compare the same machine warmed up
at different times of day (with a different mixture of traffic), the varation
is even more pronounced.

To reduce this variation, I added a new setting we can use, based on the
bytecode size of functions currently being profiled. Whenever we start
profiling a function, we add its bytecode size to a counter in ProfData. We
watch that counter to decide when to stop profiling new functions.

If you're wondering, "why not just use the amount of code in code.prof?": we're
trying to reduce variation in the *final* size of code.prof but we stop
profiling new functions long before we hit that final size. Using the amount of
bytecode eligible for profiling rather than bytes of code emitted (or even just
number of functions) provides much more stable results.

I also extended the code around Eval.NumSingleJitRequests to something that
makes sense when concurrent jitting is on. We still make sure that any threads
doing jitting are able to get the write lease for the whole request, but we
allow up to Eval.JitThreads threads to be in this state at once. This only
matters when concurrent jitting is on, which is still not yet the default.

Reviewed By: kodafb

Differential Revision: D3602569

fbshipit-source-id: 337eb879592e176cad653ee3a2a45554e05af440