Fix some stupidity and a bug in libgd
Summary:
gdImageGetTrueColorPixel applied to a truecolor image just defers to
gdImageGetPixel; so using a function pointer and calling
gdImageGetTrueColorPixel on truecolor images, and gdImageGetPixel on
indexed images doesn't make sense.

Also,

  if (ignoretransparent && c == dst->transparent) {
    gdImageSetPixel(dst, uY, (dst->sy - uX - 1), dst->transparent);
  } else {
    gdImageSetPixel(dst, uY, (dst->sy - uX - 1), c);
  }

is equivalent to

  gdImageSetPixel(dst, uY, (dst->sy - uX - 1), c);

so lets do that instead.

gdImageSelectiveBlur read cpxl from src (which could be indexed or
truecolor), and then decoded it using srcback (which is always
truecolor). If src was indexed this would produce nonsense values.
Since srcback is a copy of src, lets just read cpxl from that instead.

Reviewed By: alexmalyshev

Differential Revision: D3659016

fbshipit-source-id: 129fde47a613a314aeadfcd175711bba1af801dd