Simplify Ctx related code
Summary:
We should be able to assume that

  ActRec::m_thiz == nullptr iff ActRec::m_func->cls() == nullptr

and a lot of code already does. But it turns out there were a number
of bugs relating to pseudoMains, and a lot of cases where we were
doing unnecessary checks.

This cleans up by removing TNullptr from TCtx and dealing with the
fallout.

Reviewed By: jano

Differential Revision: D3578286

fbshipit-source-id: d7217f53a47c38d5673f7d5f64fee2c3b43046d2