Improve JobQueueDispatcher
Summary:
This does 3 things in JobQueueDispatcher and related code.
(1) make sure there is a 1-to-1 mapping of worker threads to conditon variables, to allow precise control of thread scheduling. Job reaper id is fixed. The maximum number of workers must be decided when initializing the job queue, but initial number can be lower than the maximum.
(2) threads can have 3 priority levels; the highest is not used here, it is reserved for threads with huge pages on its stack/heap
(3) use intrusive list to maintain the list of condition variables. This way, we don't need to maintain a map from conditon variable to the list it is in.

Reviewed By: markw65

Differential Revision: D3675821

fbshipit-source-id: 129762d80187a7e351240ad5f3a4a3bb5827aa04